requirements.txt
Django==4.1.3
python-dotenv
psycopg2-binary
requests
requests_oauthlib 
smtplib
